refactor: split FilesystemTools into Read/Write/Search classes (#116)
